requirements.txt
dj-database-url==0.4.1
Django==1.9.6
django-countries==3.4.1
gunicorn==19.5.0
Pillow==3.2.0
psycopg2==2.6.1
whitenoise==3.1
django-registration==2.1.2
django-custom-user==0.6
